+++ test/MANIFEST.in
