+++ test/paddlestructure.py
@@ -24,9 +24,9 @@ import numpy as np
 from pathlib import Path
 
 from ppocr.utils.logging import get_logger
-from ppstructure.predict_system import OCRSystem, save_res
-from ppstructure.table.predict_table import to_excel
-from ppstructure.utility import init_args, draw_result
+from test.predict_system import OCRSystem, save_res
+from test.table.predict_table import to_excel
+from test.utility import init_args, draw_result
 
 logger = get_logger()
 from ppocr.utils.utility import check_and_read_gif, get_image_file_list
