+++ test/predict_system.py
@@ -31,8 +31,8 @@ import layoutparser as lp
 from ppocr.utils.utility import get_image_file_list, check_and_read_gif
 from ppocr.utils.logging import get_logger
 from tools.infer.predict_system import TextSystem
-from ppstructure.table.predict_table import TableSystem, to_excel
-from ppstructure.utility import parse_args, draw_result
+from test.table.predict_table import TableSystem, to_excel
+from test.utility import parse_args, draw_result
 
 logger = get_logger()
