+++ test/setup.py
@@ -25,14 +25,14 @@ with open('../requirements.txt', encoding="utf-8-sig") as f:
 
 
 def readme():
-    with open('README_ch.md', encoding="utf-8-sig") as f:
+    with open('api_ch.md', encoding="utf-8-sig") as f:
         README = f.read()
     return README
 
 
-shutil.copytree('../ppstructure/table', './ppstructure/table')
-shutil.copyfile('../ppstructure/predict_system.py', './ppstructure/predict_system.py')
-shutil.copyfile('../ppstructure/utility.py', './ppstructure/utility.py')
+shutil.copytree('/table', './test/table')
+shutil.copyfile('/predict_system.py', './test/predict_system.py')
+shutil.copyfile('/utility.py', './test/utility.py')
 shutil.copytree('../ppocr', './ppocr')
 shutil.copytree('../tools', './tools')
 shutil.copyfile('../LICENSE', './LICENSE')
@@ -68,5 +68,5 @@ setup(
 
 shutil.rmtree('ppocr')
 shutil.rmtree('tools')
-shutil.rmtree('ppstructure')
+shutil.rmtree('test')
 os.remove('LICENSE')
