+++ test/table/predict_structure.py
@@ -32,7 +32,7 @@ from ppocr.data import create_operators, transform
 from ppocr.postprocess import build_post_process
 from ppocr.utils.logging import get_logger
 from ppocr.utils.utility import get_image_file_list, check_and_read_gif
-from ppstructure.utility import parse_args
+from test.utility import parse_args
 
 logger = get_logger()
